backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
pillow==10.2.0
huggingface-hub==0.20.3
transformers==4.37.2
python-dotenv==1.0.1